requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=./requirements.txt ./requirements.in
#
asgiref>=3
    # via flask
blinker>=1
    # via flask
click>=8
    # via flask
flask[async]>=3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==6.0.1
    # via -r requirements.in
geode-common==33.11.3
    # via
    #   -r requirements.in
    #   geode-viewables
geode-viewables==3.3.2
    # via -r requirements.in
itsdangerous>=2
    # via flask
jinja2>=3
    # via flask
markupsafe>=3
    # via
    #   flask
    #   jinja2
    #   werkzeug
opengeode-core==15.30.2
    # via
    #   -r requirements.in
    #   geode-common
    #   geode-viewables
    #   opengeode-geosciences
    #   opengeode-geosciencesio
    #   opengeode-inspector
    #   opengeode-io
opengeode-geosciences==9.5.4
    # via
    #   -r requirements.in
    #   geode-viewables
    #   opengeode-geosciencesio
opengeode-geosciencesio==5.8.6
    # via -r requirements.in
opengeode-inspector==6.8.6
    # via -r requirements.in
opengeode-io==7.4.6
    # via
    #   -r requirements.in
    #   geode-viewables
    #   opengeode-geosciencesio
werkzeug==3.1.2
    # via
    #   -r requirements.in
    #   flask
    #   flask-cors

opengeodeweb-microservice==1.*,>=1.0.9rc2
